pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "couch-control"
version = "2.0.0"
description = "Lightweight remote desktop control — control your PC from any browser"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Couch Control"}
]
keywords = ["remote-desktop", "screen-sharing", "websocket", "lightweight", "cross-platform"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
]

dependencies = [
    "aiohttp>=3.9.0",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "netifaces>=0.11.0",
]

[project.optional-dependencies]
# Fast JPEG encoding (Linux/macOS)
fast = [
    "PyTurboJPEG>=1.7.0",
]
# Windows / macOS input control (alternative to xdotool)
windows = [
    "pynput>=1.7.0",
]
# System tray icon
tray = [
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
]
# Build standalone executable
build = [
    "pyinstaller>=6.0",
]
# All optional extras
all = [
    "pynput>=1.7.0",
    "pystray>=0.19.0",
]

[project.scripts]
couch-control = "couch_control.cli:main"

[project.urls]
Homepage = "https://github.com/sithulaka/Couch-Control"
Issues   = "https://github.com/sithulaka/Couch-Control/issues"

[tool.setuptools.packages.find]
where   = ["."]
include = ["couch_control*"]

[tool.setuptools.package-data]
couch_control = ["static/*"]
